pyproject.toml
[tool.poetry]
name = "online-shop"
version = "0.1.0"
description = ""
authors = ["Valentyn Malenchak <valentyn17@gmail.com>"]
readme = "README.md"
packages = [{include = "online_shop"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
pytest = "^7.4.3"
invoke = "^2.2.0"
motor = "^3.3.2"
mypy = "^1.7.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
arrow = "^1.3.0"
httpx = "^0.25.2"
python-multipart = "^0.0.6"
types-passlib = "^1.7.7.13"
types-python-jose = "^3.3.4.8"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
